Move tests and fix paths
