--- Code/Example/feedback_on_analysis.txt
@@ -1,85 +1,0 @@
-PITCH COVERAGE =====================
-
-The pitch coverage looks good: all of the chords match the corresponding sections of the score well.
-
-
-RARE ROMAN NUMERALS =====================
-
-No rare or unconventional Roman Numerals in this analysis.
-
-
-HARMONIC RHYTHM =====================
-
-Total cases: 19
-
-In the following cases, the chord change is at an unusually weak metrical position:
-
-Measure 9, viø7 in bb appears on beat 3.5.
-
-Measure 10, V42 in Db appears on beat 3.5.
-
-Measure 10, V43 in Db appears on beat 4.5.
-
-Measure 15, viio43 in f appears on beat 2.5.
-
-Measure 16, viiø43 in Db appears on beat 2.5.
-
-Measure 18, V42 in Db appears on beat 3.5.
-
-Measure 18, V43 in Db appears on beat 4.5.
-
-Measure 25, viø7 in bb appears on beat 3.5.
-
-Measure 26, V42 in Db appears on beat 3.5.
-
-Measure 26, V43 in Db appears on beat 4.5.
-
-Measure 31, viio43 in f appears on beat 2.5.
-
-Measure 32, viiø43 in Db appears on beat 2.5.
-
-Measure 33, iiø42 in Db appears on beat 2.5.
-
-Measure 34, ii65 in Db appears on beat 1.5.
-
-Measure 34, vi64 in Db appears on beat 2.5.
-
-Measure 34, ii7 in Db appears on beat 3.5.
-
-Measure 34, V in Db appears on beat 4.5.
-
-Measure 35, I6 in Db appears on beat 1.5.
-
-Measure 35, viio7/V in Db appears on beat 2.5.
-
-
-BASS / INVERSION =====================
-
-Total cases: 4
-
-Overall bass note match: 94.51%.
-
-In these cases, the specified bass note doesn't appear in the lowest part during. (NB: pedal points are not yet supported):
-
-Measure 9, beat 1, iiø7 in bb, indicating the bass C for lowest note(s) of: ['Eb2'].
-Sorry, no inversion suggestions - none of the bass note(s) are in the chord.
-
-
-Measure 9, beat 3.5, viø7 in bb, indicating the bass G for lowest note(s) of: ['F2'].
-Sorry, no inversion suggestions - none of the bass note(s) are in the chord.
-
-
-Measure 25, beat 3.5, viø7 in bb, indicating the bass G for lowest note(s) of: ['F2'].
-Sorry, no inversion suggestions - none of the bass note(s) are in the chord.
-
-
-Measure 38, beat 1, V9 in Db, indicating the bass Ab for lowest note(s) of: ['Db3'].
-Sorry, no inversion suggestions - none of the bass note(s) are in the chord.
-
-
-
-WARNINGS =====================
-WARNING: There are 41 measures in the score but 40 in your analysis. This is usually a question of either the beginning or end: either
-1) The final chord in the analysis comes before the final measure, or
-2) There's an anacrusis in the score without an accompanying harmony (i.e. the analysis is missing measure 0). In that latter case, the score and analysis will be misaligned, and the HarmonicRanges will not work properly. Best to put in a chord of some kind for the anacrusis.
-
