--- Code/Example/profiles_by_key.arff
@@ -1,18 +1,0 @@
-@RELATION 'Chord features for profiles_by_key'
-@ATTRIBUTE 'profile' NUMERIC
-@ATTRIBUTE 'start offset' NUMERIC
-@ATTRIBUTE 'end offset' NUMERIC
-@ATTRIBUTE 'quarter length' NUMERIC
-@ATTRIBUTE 'start measure' NUMERIC
-@ATTRIBUTE 'end measure' NUMERIC
-@ATTRIBUTE 'measure length' NUMERIC
-@DATA
-"[8.5, 26.0, 0.0, 11.0, 5.0, 23.5, 11.5, 0.5, 42.0, 0.0, 13.0, 0.0]",0.0,32.0,32.0,1,8,7
-"[5.0, 4.0, 0.0, 3.0, 0.0, 9.5, 4.0, 1.0, 0.0, 3.5, 4.0, 0.0]",32.0,38.5,6.5,9,10,1
-"[5.0, 19.0, 0.0, 5.5, 5.0, 11.5, 4.5, 0.5, 23.0, 0.0, 5.0, 0.0]",38.5,57.5,19.0,10,15,5
-"[1.5, 1.5, 0.0, 0.0, 0.5, 2.0, 0.0, 0.5, 3.5, 0.0, 0.5, 0.0]",57.5,60.0,2.5,15,15,0
-"[9.5, 25.5, 0.5, 14.5, 5.0, 22.5, 16.0, 3.0, 38.0, 0.0, 16.0, 0.5]",60.0,96.0,36.0,16,24,8
-"[5.0, 4.0, 0.0, 1.5, 0.0, 11.0, 4.0, 1.0, 0.0, 3.5, 4.0, 0.0]",96.0,102.5,6.5,25,26,1
-"[5.0, 16.0, 0.0, 7.0, 5.5, 9.5, 6.0, 2.5, 23.0, 0.0, 5.0, 0.0]",102.5,121.5,19.0,26,31,5
-"[1.5, 1.5, 0.0, 0.0, 0.5, 2.0, 0.0, 0.5, 3.5, 0.0, 0.5, 0.0]",121.5,124.0,2.5,31,31,0
-"[11.5, 47.0, 1.0, 18.5, 1.5, 21.0, 20.0, 1.5, 29.0, 6.5, 10.5, 0.5]",124.0,164.0,40.0,32,41,9
