--- Code/Example/profiles_by_measure.tsv
@@ -1,42 +1,0 @@
-measure	profile
-1	[1.0, 3.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-2	[1.0, 3.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-3	[1.0, 3.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-4	[1.0, 4.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-5	[1.5, 6.0, 0, 0, 2.5, 2.0, 0, 0, 5.0, 0, 0.5, 0]
-6	[1.0, 2.0, 0, 1.0, 2.5, 0, 0, 0.5, 3.0, 0, 2.0, 0]
-7	[1.0, 1.0, 0, 2.0, 0, 2.0, 3.5, 0, 2.0, 0, 7.5, 0]
-8	[1.0, 2.0, 0, 2.0, 0, 3.5, 2.0, 0, 8.0, 0, 1.0, 0]
-9	[2.0, 4.0, 0, 3.0, 0, 6.0, 4.0, 1.0, 0, 1.0, 4.0, 0]
-10	[4.0, 0.5, 0, 1.0, 0, 4.5, 1.0, 0, 3.0, 2.5, 0, 0]
-11	[1.0, 3.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-12	[1.0, 4.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-13	[1.0, 6.5, 0, 0, 2.5, 2.0, 0, 0, 5.0, 0, 0.5, 0]
-14	[1.0, 2.0, 0, 1.0, 2.5, 0, 0, 0.5, 3.0, 0, 2.0, 0]
-15	[1.5, 3.5, 0, 0.5, 0.5, 2.5, 0.5, 0.5, 3.5, 0, 2.0, 0]
-16	[0.5, 2.5, 0.5, 0.5, 0, 3.5, 2.0, 0, 1.0, 0, 3.0, 0.5]
-17	[2.0, 1.5, 0, 4.5, 0, 0.5, 4.0, 0, 4.0, 0, 1.0, 0]
-18	[1.5, 3.0, 0, 1.0, 0, 3.0, 1.5, 0.5, 3.0, 0, 0, 0]
-19	[1.0, 3.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-20	[1.0, 4.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-21	[0.5, 5.5, 0, 0, 2.5, 2.0, 0, 0, 5.0, 0, 0.5, 0]
-22	[1.0, 2.0, 0, 1.5, 2.5, 0, 0, 2.5, 3.0, 0, 2.0, 0]
-23	[1.0, 1.0, 0, 2.0, 0, 2.0, 3.5, 0, 2.0, 0, 7.5, 0]
-24	[1.0, 2.0, 0, 2.0, 0, 3.5, 2.0, 0, 8.0, 0, 1.0, 0]
-25	[2.0, 4.0, 0, 1.5, 0, 6.0, 4.0, 1.0, 0, 1.0, 4.0, 0]
-26	[4.0, 0.5, 0, 1.5, 0, 6.5, 1.5, 0, 2.5, 2.5, 0, 0]
-27	[1.0, 2.5, 0, 2.0, 0, 3.5, 2.5, 0, 7.5, 0, 0.5, 0]
-28	[1.0, 2.5, 0, 1.5, 0, 2.0, 1.5, 0, 5.0, 0, 0.5, 0]
-29	[1.0, 6.5, 0, 0, 3.0, 2.0, 0, 0, 5.0, 0, 0.5, 0]
-30	[1.0, 2.0, 0, 1.5, 2.5, 0, 0, 2.5, 3.0, 0, 2.0, 0]
-31	[1.5, 3.5, 0, 0.5, 0.5, 2.5, 0.5, 0.5, 3.5, 0, 2.0, 0]
-32	[0.5, 2.5, 0.5, 0.5, 0, 3.5, 2.0, 0, 1.5, 0, 3.5, 0.5]
-33	[1.0, 3.0, 0.5, 1.5, 0, 1.0, 3.0, 0, 2.0, 1.5, 0, 0]
-34	[0, 7.0, 0, 2.0, 0, 2.0, 3.0, 0, 1.0, 0, 3.0, 0]
-35	[3.0, 4.0, 0, 2.0, 0.5, 0.5, 2.5, 0.5, 6.0, 3.0, 1.5, 0]
-36	[1.0, 6.5, 0, 1.5, 0, 4.0, 1.5, 0, 6.0, 0, 0.5, 0]
-37	[1.5, 3.5, 0, 2.5, 1.0, 1.5, 3.5, 1.0, 1.5, 0.5, 1.5, 0]
-38	[4.0, 4.0, 0, 8.0, 0, 0.5, 4.5, 0, 1.5, 0.5, 0.5, 0]
-39	[0.5, 0.5, 0, 0.5, 0, 0, 0, 0, 1.5, 1.0, 0, 0]
-40	[0, 8.0, 0, 0, 0, 4.0, 0, 0, 4.0, 0, 0, 0]
-41	[0, 8.0, 0, 0, 0, 4.0, 0, 0, 4.0, 0, 0, 0]
