+++ Code/Pitch_profiles/chord_features.py
@@ -31,17 +31,15 @@ TODO: Currently limited to single chord. expand to – chord progressions
 
 """
 
-import chord_comparisons
-import chord_profiles
-import chord_usage
-import normalisation_comparison
+from . import chord_comparisons
+from . import chord_profiles
+from . import chord_usage
+from . import normalisation_comparison
 
 from music21 import analysis
 from music21 import roman
 
 from typing import Union, Optional
-
-import unittest
 
 
 # ------------------------------------------------------------------------------
@@ -367,58 +365,3 @@ def getCommonPercentage(thisDict, thisKey):
         return 0
 
     return round(thisPercent / maxPercent, 3)
-
-
-# ------------------------------------------------------------------------------
-
-class Test(unittest.TestCase):
-
-    def testSingleChordFeatures(self):
-        rn = roman.RomanNumeral('I')  # Tonic major
-        profile = [309.17, 37.31, 10.09, 376.62, 10.75, 21.43,
-                   316.04, 1.02, 19.79, 33.68, 24.33, 22.97]  # c diminished
-        testFeaturesSet = SingleChordFeatures(rn, profile)
-
-        for pair in (
-                ('chordQualityVector',
-                 [0, 0, 1, 0, 0, 0, 0, 0, 0, 0]),  # major
-                ('thirdTypeVector',
-                 [0, 1, 0]),  # major
-                ('fifthTypeVector',
-                 [0, 1, 0, 0]),  # perfect
-                ('seventhTypeVector',
-                 [0, 0, 0, 1]),  # None
-                ('rootPitchClassVector',
-                 [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]),  # C
-                ('hauptFunctionVector',
-                 [1, 0, 0, 0, 0, 0, 0]),  # T
-                ('functionVector',
-                 [1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]),  # T
-                ('chosenChordPCPVector',
-                 [1, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0]),  # C major
-                ('bestFitChordPCPVector',
-                 [1, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0, 0]),  # c diminished
-
-                ('chordTypeMatchVector',  # False, major vs dim.
-                 [0]),
-                ('chordRotationMatchVector',  # True, C = C
-                 [1]),
-
-                ('distanceToBestFitChordPCPVector',
-                 [0.307]),
-                ('distanceToChosenChordVector',
-                 [0.7285]),
-
-                ('fullChordCommonnessVector',
-                 [1]),  # Most common = 1
-                ('simplifiedChordCommonnessVector',
-                 [0.908])  # Still very common
-        ):
-            vector = getattr(testFeaturesSet, pair[0])
-            self.assertEqual(vector, pair[1])
-
-
-# ------------------------------------------------------------------------------
-
-if __name__ == '__main__':
-    unittest.main()
