+++ Code/Pitch_profiles/get_distributions.py
@@ -46,12 +46,11 @@ and timing information as appropriate to the data source with start, end and len
 
 # ------------------------------------------------------------------------------
 
-import unittest
 import os
 import csv
 
-import normalisation_comparison
-import chord_features
+from . import normalisation_comparison
+from . import chord_features
 
 from typing import Optional
 
@@ -571,47 +570,3 @@ class DistributionsFromTabular:
                                                       round_places=self.round_places)
         else:
             return [round(x, self.round_places) for x in dist]
-
-
-# ------------------------------------------------------------------------------
-
-class Test(unittest.TestCase):
-
-    def test_two_songs(self):
-        base_path = '../../Corpus/OpenScore-LiederCorpus/'
-
-        p = base_path + 'Reichardt,_Louise/Zwölf_Gesänge,_Op.3/01_Frühlingsblumen/'
-        test_p = DistributionsFromTabular(path_to_tab=p + 'slices_with_analysis.tsv')
-        test_p.get_profiles_by_key()
-        self.assertEqual(len(test_p.profiles_by_key), 1)  # i.e. only one key
-
-        q = base_path + 'Schumann,_Clara/Lieder,_Op.12/04_Liebst_du_um_Schönheit/'
-        test_q = DistributionsFromTabular(path_to_tab=q + 'slices_with_analysis.tsv')
-        test_q.get_profiles_by_key()
-        self.assertEqual(len(test_q.profiles_by_key), 9)  # i.e. 9 local key areas
-        self.assertEqual(test_q.profiles_by_key[0]['key'], 'Db')  # Starting in Db
-        self.assertEqual(test_q.profiles_by_key[-1]['key'], 'Db')  # And ending there too
-
-        new_path = '../Example/'
-        test_r = DistributionsFromTabular(path_to_tab=new_path + 'slices_with_analysis.tsv')
-        self.assertEqual(test_q.get_profiles_by_key(),
-                         test_r.get_profiles_by_key())  # i.e. exact copy of corpus entry
-
-        test_s = DistributionsFromTabular(path_to_tab=new_path + 'slices_with_analysis.tsv',
-                                          include_features=True)
-
-        # NB: WRITES
-        valid_formats = ['.csv', '.tsv', '.arff', '.json']
-        valid_by_what = ['key', 'chord', 'measure']
-        for vf in valid_formats:
-            for vbw in valid_by_what:
-                for this_bool in [False, True]:
-                    test_r.write_distributions(by_what=vbw,
-                                               write_features=this_bool,
-                                               out_format=vf)
-
-
-# ------------------------------------------------------------------------------
-
-if __name__ == '__main__':
-    unittest.main()
