+++ Code/__init__.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+CODE_FOLDER = Path(__file__).parent
+REPO_FOLDER = CODE_FOLDER.parent
+CORPUS_FOLDER = REPO_FOLDER / "Corpus"
