+++ Tests/Resources/Example/score.mxl
