+++ Tests/Resources/Example/template.txt
