+++ Tests/Resources/testPartialAnalysis.mxl
