+++ Tests/__init__.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+TEST_FOLDER = Path(__file__).parent
+TEST_RESOURCES_FOLDER = TEST_FOLDER / "Resources"
